Make UseStatusCodePagesWithReExecute clear Endpoints #13715
